requirements.txt
google-cloud-aiplatform>=1.38.1
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
vertexai>=0.0.1
structlog>=24.1.0
slack-bolt>=1.18.0  # For future Slack integration
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.1.1 